refactor: move adapters out of dependency package
