+++ cmd/glasskube/cmd/install.go
@@ -7,9 +7,9 @@ import (
 
 	"github.com/fatih/color"
 	"github.com/glasskube/glasskube/api/v1alpha1"
+	clientadapter "github.com/glasskube/glasskube/internal/adapter/goclient"
 	"github.com/glasskube/glasskube/internal/cliutils"
 	"github.com/glasskube/glasskube/internal/dependency"
-	clientadapter "github.com/glasskube/glasskube/internal/dependency/adapter/goclient"
 	"github.com/glasskube/glasskube/internal/repo"
 	"github.com/glasskube/glasskube/pkg/client"
 	"github.com/glasskube/glasskube/pkg/condition"
