+++ internal/adapter/adapter.go
