+++ internal/adapter/goclient/goclientadapter.go
