+++ internal/webhook/package_validator_test.go
@@ -5,9 +5,9 @@ import (
 	"time"
 
 	"github.com/glasskube/glasskube/api/v1alpha1"
+	ctrladapter "github.com/glasskube/glasskube/internal/adapter/controllerruntime"
 	"github.com/glasskube/glasskube/internal/controller/owners"
 	"github.com/glasskube/glasskube/internal/dependency"
-	ctrladapter "github.com/glasskube/glasskube/internal/dependency/adapter/controllerruntime"
 	fakerepo "github.com/glasskube/glasskube/internal/repo/client/fake"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
